+++ src/components/dropdown/dropdown.ts
@@ -31,7 +31,6 @@ import {
 } from '../common/mixins/combo-box.js';
 import type { AbstractConstructor } from '../common/mixins/constructor.js';
 import { EventEmitterMixin } from '../common/mixins/event-emitter.js';
-import { SizableMixin } from '../common/mixins/sizable.js';
 import {
   findElementFromEventPath,
   getElementByIdFromRoot,
@@ -73,12 +72,10 @@ export interface IgcDropdownEventMap {
 @blazorAdditionalDependencies(
   'IgcDropdownItemComponent, IgcDropdownHeaderComponent, IgcDropdownGroupComponent'
 )
-export default class IgcDropdownComponent extends SizableMixin(
-  EventEmitterMixin<
-    IgcDropdownEventMap,
-    AbstractConstructor<IgcBaseComboBoxLikeComponent>
-  >(IgcBaseComboBoxLikeComponent)
-) {
+export default class IgcDropdownComponent extends EventEmitterMixin<
+  IgcDropdownEventMap,
+  AbstractConstructor<IgcBaseComboBoxLikeComponent>
+>(IgcBaseComboBoxLikeComponent) {
   public static readonly tagName = 'igc-dropdown';
   public static styles = [styles, shared];
 
@@ -230,8 +227,6 @@ export default class IgcDropdownComponent extends SizableMixin(
       .set(enterKey, this.onEnterKey)
       .set(homeKey, this.onHomeKey)
       .set(endKey, this.onEndKey);
-
-    this.size = 'medium';
   }
 
   protected override async firstUpdated() {